reduces transaction time from 25s to 6s
